tests: refactor cli tests, move test specific fixtures to tests file
Changelog: None
